+ Cast block with busy state
