+++ go.mod
@@ -1,5 +1,3 @@
 module rhymald/mag-delta
 
 go 1.18
-
-require golang.org/x/crypto v0.5.0
